requirements.txt
discord.py[voice]>=2.7.1
yt-dlp>=2024.12.0
spotipy>=2.24.0
python-dotenv>=1.0.0
quart>=0.19.0
hypercorn>=0.17.0
aiosqlite>=0.20.0
gTTS>=2.5.0
